SVM: Loader: absorb rest of `account_loader`
